backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.2
redis==5.0.8
celery==5.4.0
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
pydantic-settings==2.4.0
structlog==24.4.0
httpx==0.27.2
python-multipart==0.0.9
email-validator==2.2.0
ruff==0.13.2
